+++ lumin/optimisation/hyper_param.py
@@ -1,4 +1,4 @@
-from typing import Tuple, Dict, List, Optional, Union
+from typing import Tuple, Dict, List, Optional, Union, Callable, Generator
 from fastprogress import master_bar, progress_bar
 import numpy as np
 from collections import OrderedDict
@@ -10,13 +10,14 @@ from sklearn.ensemble import RandomForestRegressor, RandomForestClassifier
 from ..nn.data.fold_yielder import FoldYielder
 from ..nn.data.batch_yielder import BatchYielder
 from ..nn.models.model_builder import ModelBuilder
-from ..nn.models.model import Model
-from ..nn.callbacks.opt_callbacks import LRFinder
+from ..nn.models.model import Model, OldModel
+from ..nn.callbacks.opt_callbacks import LRFinder, OldLRFinder
 from ..nn.callbacks.cyclic_callbacks import AbsCyclicCallback
 from ..nn.callbacks.model_callbacks import AbsModelCallback
-
 from ..plotting.training import plot_lr_finders
 from ..plotting.plot_settings import PlotSettings
+
+from torch import optim, Tensor
 
 import matplotlib.pyplot as plt
 
@@ -76,31 +77,17 @@ def get_opt_rf_params(x_trn:np.ndarray, y_trn:np.ndarray, x_val:np.ndarray, y_va
     return best_params, best_m
 
 
-def fold_lr_find(fy:FoldYielder, model_builder:ModelBuilder, bs:int,
-                 train_on_weights:bool=True, shuffle_fold:bool=True, n_folds:int=-1, lr_bounds:Tuple[float,float]=[1e-5, 10],
-                 callback_partials:Optional[List[partial]]=None, plot_settings:PlotSettings=PlotSettings(),
-                 bulk_move:bool=True, plot_savename:Optional[str]=None) -> List[LRFinder]:
+def old_fold_lr_find(fy:FoldYielder, model_builder:ModelBuilder, bs:int,
+                     train_on_weights:bool=True, shuffle_fold:bool=True, n_folds:int=-1, lr_bounds:Tuple[float,float]=[1e-5, 10],
+                     callback_partials:Optional[List[partial]]=None, plot_settings:PlotSettings=PlotSettings(),
+                     bulk_move:bool=True, plot_savename:Optional[str]=None) -> List[OldLRFinder]:
     r'''
-    Wrapper function for training using :class:`~lumin.nn.callbacks.opt_callbacks.LRFinder` which runs a Smith LR range test (https://arxiv.org/abs/1803.09820)
-    using folds in :class:`~lumin.nn.data.fold_yielder.FoldYielder`.
-    Trains models for 1 fold, interpolating LR between set bounds. This repeats for each fold in :class:`~lumin.nn.data.fold_yielder.FoldYielder`, and loss
-    evolution is averaged.
-
-    Arguments:
-        fy: :class:`~lumin.nn.data.fold_yielder.FoldYielder` providing training data
-        model_builder: :class:`~lumin.nn.models.model_builder.ModelBuilder` providing networks and optimisers
-        bs: batch size
-        train_on_weights: If weights are present, whether to use them for training
-        shuffle_fold: whether to shuffle data in folds
-        n_folds: if >= 1, will only train n_folds number of models, otherwise will train one model per fold
-        lr_bounds: starting and ending LR values
-        callback_partials: optional list of functools.partial, each of which will a instantiate :class:`~lumin.nn.callbacks.callback.Callback` when called        
-        plot_settings: :class:`~lumin.plotting.plot_settings.PlotSettings` class to control figure appearance
-        savename: Optional name of file to which to save the plot
-
-    Returns:
-        List of :class:`~lumin.nn.callbacks.opt_callbacks.LRFinder` which were used for each model trained
+    .. Attention:: This class is depreciated in favour of :class:`~lumin.nn.callbacks.data_callbacks.fold_lr_find`.
+        It is a copy of the old `fold_lr_find` class used in lumin<=0.6.
+        It will be removed in V0.8
     '''
+
+    # XXX remove in V0.8
 
     if callback_partials is None: callback_partials = []
     idxs = range(fy.n_folds) if n_folds < 1 else range(min(n_folds, fy.n_folds))
@@ -108,7 +95,7 @@ def fold_lr_find(fy:FoldYielder, model_builder:ModelBuilder, bs:int,
     tmr = timeit.default_timer()
     nb = None
     for trn_id in progress_bar(idxs):
-        model = Model(model_builder)
+        model = OldModel(model_builder)
         trn_fold = fy.get_fold(trn_id)
         if nb is None: nb = len(trn_fold['targets'])//bs
         lr_finder = LRFinder(nb=nb, lr_bounds=lr_bounds, model=model)
@@ -126,6 +113,55 @@ def fold_lr_find(fy:FoldYielder, model_builder:ModelBuilder, bs:int,
         model.fit(batch_yielder, callbacks+[lr_finder])
         lr_finders.append(lr_finder)
     del batch_yielder
+    del model
+        
+    print("LR finder took {:.3f}s ".format(timeit.default_timer()-tmr))
+    plot_lr_finders(lr_finders, loss_range='auto', settings=plot_settings, log_y='auto' if 'regress' in model_builder.objective.lower() else False,
+                    savename=plot_savename)
+    return lr_finders
+
+
+def fold_lr_find(fy:FoldYielder, model_builder:ModelBuilder, bs:int, n_epochs:int=1,
+                 train_on_weights:bool=True, n_folds:int=-1, lr_bounds:Tuple[float,float]=[1e-5, 10],
+                 cb_partials:Optional[List[partial]]=None, plot_settings:PlotSettings=PlotSettings(),
+                 bulk_move:bool=True, plot_savename:Optional[str]=None, opt:Optional[Callable[[Generator],optim.Optimizer]]=None,
+                 loss:Optional[Callable[[],Callable[[Tensor,Tensor],Tensor]]]=None) -> List[LRFinder]:
+    r'''
+    Wrapper function for training using :class:`~lumin.nn.callbacks.opt_callbacks.LRFinder` which runs a Smith LR range test (https://arxiv.org/abs/1803.09820)
+    using folds in :class:`~lumin.nn.data.fold_yielder.FoldYielder`.
+    Trains models for a set number of fold, interpolating LR between set bounds. This repeats for each fold in :class:`~lumin.nn.data.fold_yielder.FoldYielder`,
+    and loss evolution is averaged.
+
+    Arguments:
+        fy: :class:`~lumin.nn.data.fold_yielder.FoldYielder` providing training data
+        model_builder: :class:`~lumin.nn.models.model_builder.ModelBuilder` providing networks and optimisers
+        bs: batch size
+        n_epochs: number of epochs to train per fold
+        train_on_weights: If weights are present, whether to use them for training
+        shuffle_fold: whether to shuffle data in folds
+        n_folds: if >= 1, will only train n_folds number of models, otherwise will train one model per fold
+        lr_bounds: starting and ending LR values
+        cb_partials: optional list of functools.partial, each of which will a instantiate :class:`~lumin.nn.callbacks.callback.Callback` when called        
+        plot_settings: :class:`~lumin.plotting.plot_settings.PlotSettings` class to control figure appearance
+        savename: Optional name of file to which to save the plot
+
+    Returns:
+        List of :class:`~lumin.nn.callbacks.opt_callbacks.LRFinder` which were used for each model trained
+    '''
+
+    if cb_partials is None: cb_partials = []
+    idxs = range(fy.n_folds) if n_folds < 1 else range(min(n_folds, fy.n_folds))
+    lr_finders,nb = [],None
+    tmr = timeit.default_timer()
+    for trn_idx in progress_bar(idxs):
+        model = Model(model_builder)
+        cbs = []
+        for c in cb_partials: cbs.append(c(model=model))
+        lrf = LRFinder(lr_bounds=lr_bounds, nb=nb, model=model)
+        model.fit(n_epochs=n_epochs, fy=fy, bs=bs, bulk_move=bulk_move, train_on_weights=train_on_weights, trn_idxs=[trn_idx], cbs=cbs+[lrf], opt=opt,
+                  loss=loss)
+        if nb is None: nb = lrf.nb  # Ensure all LR FInders follow same LR history
+        lr_finders.append(lrf)
     del model
         
     print("LR finder took {:.3f}s ".format(timeit.default_timer()-tmr))